+++ packages/app/cypress/support/index.js
@@ -1,6 +1,0 @@
-/*
-* cypress/support/index.js
-*
-* Collects together custom Cypress commands.
-*/
-import './auth'