+++ tools/npm-version-check.sh
@@ -6,14 +6,18 @@ set -eu -o pipefail
 # We are not _really_ requiring npm >= 7.7.0 but there were some bugs in earlier 7-series, and npm 6 seems incompatible
 # (and slow), so for the sake of keeping moving parts down, 7.7.0 or later. Others - at your own risk!!
 #
-
-VER="$(npm --version)"
-  # e.g. "7.9.0"
-
-# When reaching npm 10; remove the whole checks
+# Requires:
+#   - sort -C -V   (macOS ok)
 #
-if [[ "$VER" < "7.7" ]]; then
-  >&2 echo "WARNING: Your npm is less than recommended version ($VER < 7.7.0). Proceed with caution or upgrade!\n"
-fi
+
+_HAVE="$(npm --version)"
+  # e.g. "7.9.0"
+_WANT="7.7"
+
+# Check the input is sorted ('sort -C' return code is non-zero if unsorted)
+#
+printf "%s\n" $_WANT $_HAVE | sort -C -V || (
+  >&2 echo "WARNING: Your npm is less than recommended version ($_HAVE < $_WANT). Proceed with caution or upgrade!\n"
+)
 
 # peaceful exit